Capstone Full Upload
This is a full upload of the original and updated code for all three artifacts of my capstone project for SNHU. This also contains Github Pages files creating an ePortfolio of this work. A second commit will be made updating the links to the repo code currently containing a placeholder following this commit.